fixed nits for msvs compiler
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1696 d60aafaf-7936-0410-ba4d-d307febf7868

+++ src/Preferences.cpp
@@ -76,7 +76,7 @@ void Preferences::setIntProp(String name, int val)
 
 // String Props
 
-float Preferences::getDoubleProp(String name, double def)
+double Preferences::getDoubleProp(String name, double def)
 {
   return props->getDoubleValue(name, def);
 }
